pyproject.toml
[build-system]
requires = ["setuptools>=80.9.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "agent-utilities"
version = "0.2.40"
description = "Agent Utilities for Pydantic AI Agents"
readme = "README.md"
classifiers = [
    "Development Status :: 4 - Beta",
    "License :: OSI Approved :: MIT License",
    "Environment :: Console",
    "Operating System :: POSIX :: Linux",
    "Programming Language :: Python :: 3"]
requires-python = ">=3.11,<3.14"
dependencies = [
    "tree-sitter>=0.23.2",
    "requests>=2.32.5",
    "urllib3>=2.3.0",
    "pydantic[email]>=2.8.2",
    "pydantic-settings>=2.0.0",
    "PyYAML>=6.0.0",
    "python-dotenv>=1.2.1",
    "networkx>=3.0",
    "ladybug>=0.15.3",
    "numpy",
    "scipy",
    "pathspec"]

[[project.authors]]
name = "Audel Rouhi"
email = "knucklessg1@gmail.com"

[project.license]
text = "MIT"

[project.optional-dependencies]
mcp = [
    "fastmcp>=3.2.0",
    "eunomia-mcp>=0.3.10",
    "fastapi>=0.131.0",
    "httpx>=0.28.1"
]
agent = [
    "agent-utilities[mcp,ag-ui,terminal]>=0.2.40",
    "universal-skills>=0.1.56",
    "skill-graphs>=0.1.38",
    "pydantic-ai-skills>=0.4.1",
    "pydantic-ai-slim[fastmcp,openai,a2a,ag-ui,web,cli]>=1.73.0,<2.0.0",
    "pydantic-graph>=0.1.8"
]
agent-google = [
    "pydantic-ai-slim[google]>=1.73.0,<2.0.0"
]
agent-groq = [
    "pydantic-ai-slim[groq]>=1.73.0,<2.0.0"
]
agent-mistral = [
    "pydantic-ai-slim[mistral]>=1.73.0,<2.0.0"
]
agent-anthropic = [
    "pydantic-ai-slim[anthropic]>=1.73.0,<2.0.0"
]
agent-huggingface = [
    "pydantic-ai-slim[huggingface]>=1.73.0,<2.0.0"
]
ag-ui = [
    "agent-webui>=0.1.32"]
terminal = [
    "agent-terminal-ui>=0.1.35"]
browser = [
    "playwright>=1.49.1"]
acp = [
    "pydantic-acp>=0.1.0",
    "acpkit[pydantic]>=0.1.0"
]
embeddings = [
    "llama-index-core>=0.14.15"]
embeddings-openai = [
    "agent-utilities[embeddings]>=0.2.40",
    "llama-index-embeddings-openai>=0.5.1"]
embeddings-huggingface = [
    "agent-utilities[embeddings]>=0.2.40",
    "llama-index-embeddings-huggingface>=0.6.1"]
embeddings-ollama = [
    "agent-utilities[embeddings]>=0.2.40",
    "llama-index-embeddings-ollama>=0.8.6"]
logfire = [
    "agent-utilities[agent]>=0.2.40",
    "pydantic-ai-slim[logfire]>=1.73.0,<2.0.0"
]
graph = [
    "networkx>=3.0",
    "ladybug>=0.15.3",
    "tree-sitter-python>=0.23.0",
    "tree-sitter-javascript>=0.23.0",
    "tree-sitter-typescript>=0.23.0"
]
neo4j = [
    "neo4j>=5.14.1"
]
falkordb = [
    "falkordb>=1.0.1"
]
owl = [
    "owlready2>=0.46",
    "rdflib>=7.0.0"
]
stardog = [
    "agent-utilities[owl]",
    "pystardog>=0.20.0"
]
test = [
    "hypothesis>=6.152.1",
    "pytest-cov",
    "mypy",
    "ruff",
    "mutmut",
    "pytest-asyncio",
    "pytest-env",
    "opentelemetry-instrumentation-starlette",
    "opentelemetry-instrumentation-fastapi",
    "opentelemetry-instrumentation-asgi",
    "repository-manager",
    "agent-utilities[agent,graph,acp,owl]"
]
all = [
    "agent-utilities[mcp,agent,agent-google,agent-groq,agent-mistral,agent-anthropic,agent-huggingface,embeddings-openai,embeddings-huggingface,embeddings-ollama,logfire,acp,graph,browser,test,owl,stardog]>=0.2.40"]

[tool.setuptools]
include-package-data = true

[tool.setuptools.package-data]
agent_utilities = ["agent_data/**", "prompts/*.md"]

[tool.setuptools.packages.find]
where = ["."]

[tool.ruff]
line-length = 88
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B"]
ignore = ["E402", "B008", "E501", "B007"]

[tool.mypy]
python_version = "3.11"
ignore_missing_imports = true
check_untyped_defs = true
disallow_untyped_defs = false

[dependency-groups]

[tool.uv.sources]
agent-terminal-ui = { path = "../agent-terminal-ui" }
agent-webui = { path = "../agent-webui" }
repository-manager = { path = "../agents/repository-manager", editable = true }
universal-skills = { path = "../skills/universal-skills" }
skill-graphs = { path = "../skills/skill-graphs" }
acpkit = { path = "../../open-source-libraries/acpkit" }
[tool.pytest.ini_options]
asyncio_mode = "auto"
pythonpath = "."
testpaths = ["tests/unit", "tests/integration"]
markers = [
    "integration: Integration tests using subprocess/TestClient (no external services)",
    "live: Tests requiring live external services (LLM endpoints, real networks)"
]
env = [
    "AGENT_UTILITIES_TESTING=true",
    "VALIDATION_MODE=true"
]

[tool.coverage.run]
source = ["agent_utilities"]
omit = ["tests/*", "setup.py"]

[tool.coverage.report]
show_missing = true
fail_under = 70
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "if self.debug:",
    "if __name__ == .__main__.:",
    "raise AssertionError",
    "raise NotImplementedError",
    "if TYPE_CHECKING:",
    "pass",
    "except ImportError:"
]
